# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "description": "Creates a new user account in the system. This endpoint is publicly accessible.",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data - Check the request body format"
          },
          "409": {
            "description": "Email already exists - The provided email is already registered"
          }
        },
        "summary": "Register new user",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "description": "Authenticates a user and returns a JWT token for subsequent API calls",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "User credentials for authentication",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful - Returns JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials - Email or password is incorrect"
          }
        },
        "summary": "User login",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/profile": {
      "get": {
        "description": "Retrieves the profile information of the currently authenticated user",
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or expired token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user profile",
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/api/projects/{projectId}/cultural-data": {
      "post": {
        "operationId": "CulturalDataController_create",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCulturalDataDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cultural data created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CulturalDataResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data."
          },
          "404": {
            "description": "Project or referenced Region not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Creates new cultural data within a project",
        "tags": [
          "Cultural Data"
        ]
      },
      "get": {
        "operationId": "CulturalDataController_findAll",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of cultural data.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CulturalDataResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Project not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Gets all cultural data for a project",
        "tags": [
          "Cultural Data"
        ]
      }
    },
    "/api/projects/{projectId}/cultural-data/{culturalDataId}": {
      "get": {
        "operationId": "CulturalDataController_findOne",
        "parameters": [
          {
            "name": "culturalDataId",
            "required": true,
            "in": "path",
            "description": "Key of the cultural data (e.g., direct-and-formal)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cultural data found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CulturalDataResponse"
                }
              }
            }
          },
          "404": {
            "description": "Project or Cultural data not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Gets cultural data by ID within a project",
        "tags": [
          "Cultural Data"
        ]
      },
      "patch": {
        "operationId": "CulturalDataController_update",
        "parameters": [
          {
            "name": "culturalDataId",
            "required": true,
            "in": "path",
            "description": "Key of the cultural data to update (e.g., direct-and-formal)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCulturalDataDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cultural data updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CulturalDataResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data."
          },
          "404": {
            "description": "Project or Cultural data not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Updates cultural data by ID within a project",
        "tags": [
          "Cultural Data"
        ]
      },
      "delete": {
        "operationId": "CulturalDataController_remove",
        "parameters": [
          {
            "name": "culturalDataId",
            "required": true,
            "in": "path",
            "description": "Key of the cultural data to delete (e.g., direct-and-formal)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cultural data deleted."
          },
          "404": {
            "description": "Project or Cultural data not found."
          },
          "409": {
            "description": "Conflict on delete (referenced by other entities)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deletes cultural data by ID within a project",
        "tags": [
          "Cultural Data"
        ]
      }
    },
...
  },
  "info": {
    "title": "japm.app API",
    "description": "API for the japm.app Prompt Engineering application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "User email address",
            "example": "test@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "description": "User password",
            "example": "password123"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "description": "User full name",
            "example": "Test User"
          }
        },
        "required": [
          "email",
          "password",
          "name"
        ]
      },
      "UserProfileResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique user identifier",
            "example": "clg123xyz"
          },
          "email": {
            "type": "string",
            "description": "User's email address",
            "example": "user@example.com"
          },
          "name": {
            "type": "string",
            "description": "User's full name",
            "example": "John Doe"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Account creation timestamp",
            "example": "2024-03-20T10:30:00Z"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "createdAt"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "User email address",
            "example": "test@example.com"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "password123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "JWT access token for authentication",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "access_token"
        ]
      },
      "CreateRegionDto": {
        "type": "object",
        "properties": {
          "languageCode": {
            "type": "string",
            "minLength": 5,
            "maxLength": 5,
            "description": "Unique language code acting as ID",
            "example": "de-DE"
          },
          "name": {
            "type": "string",
            "description": "Name of the region",
            "example": "Germany"
          },
          "parentRegionId": {
            "type": "string",
            "description": "languageCode of the parent region (optional)",
            "example": "eu"
          },
          "timeZone": {
            "type": "string",
            "description": "Time zone",
            "example": "Europe/Berlin"
          },
          "notes": {
            "type": "string",
            "description": "Default formality level (optional)",
            "example": "Formal"
          }
        },
        "required": [
          "languageCode",
          "name"
        ]
      },
...
      "CreateCulturalDataDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "description": "Unique key for the cultural data within the project",
            "example": "direct-and-formal"
          },
          "regionId": {
            "type": "string",
            "minLength": 5,
            "maxLength": 5,
            "description": "ID (CUID) of the Region this data applies to"
          },
          "style": {
            "type": "string",
            "description": "The style of communication"
          },
          "notes": {
            "type": "string",
            "description": "Additional notes"
          }
        },
        "required": [
          "key",
          "regionId",
          "style"
        ]
      },
      "CulturalDataResponse": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "description": "Unique key for the cultural data within the project",
            "example": "direct-and-formal"
          },
          "regionId": {
            "type": "string",
            "minLength": 5,
            "maxLength": 5,
            "description": "ID (CUID) of the Region this data applies to"
          },
          "style": {
            "type": "string",
            "description": "The style of communication"
          },
          "notes": {
            "type": "string",
            "description": "Additional notes"
          },
          "region": {
            "$ref": "#/components/schemas/CreateRegionDto"
          },
          "projectId": {
            "type": "string"
          }
        },
        "required": [
          "key",
          "regionId",
          "style",
          "region",
          "projectId"
        ]
      },
      "UpdateCulturalDataDto": {
        "type": "object",
        "properties": {
          "style": {
            "type": "string",
            "description": "The style of communication"
          },
          "notes": {
            "type": "string",
            "description": "Additional notes"
          },
          "key": {
            "type": "string",
            "description": "The key identifier for the cultural data"
          }
        }
      },
...
    }
  }
}